refactor: substitui TableContainer por PaginatedTable e ajusta lógica de dados
