--- src/features/acf/modules/AcfDashboardPage/modules/PanelSelector/modules/dashboards/modules/table/Table.container.tsx
@@ -1,28 +1,0 @@
-import { PaginatedTable } from "./Table.presentation";
-import type { GridPaginationModel } from "@mui/x-data-grid";
-import type { AcfDashboardType } from "@/features/acf/modules/AcfDashboardPage/types";
-import { tableDataHook } from "./Table.hook";
-import { useState } from "react";
-
-type TableContainerProps = {
-    acfDashboardType: AcfDashboardType;
-};
-
-export const TableContainer = ({ 
-    acfDashboardType 
-}: TableContainerProps
-) => {
-    const [paginationModel, setPaginationModel] = useState<GridPaginationModel>({
-        page: 0,
-        pageSize: 8,
-    });
-    const { response, isLoading } = tableDataHook(acfDashboardType, paginationModel);
-    return (
-        <PaginatedTable
-            response={response}
-            paginationModel={paginationModel}
-            setPaginationModel={setPaginationModel}
-            isLoading={isLoading}
-        />
-    );
-}
